requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv
pydantic==2.9.2
pydantic-settings==2.4.0
httpx==0.27.2
python-jose[cryptography]==3.3.0
pyjwt==2.9.0
cachetools==5.5.0

SQLAlchemy==2.0.34
asyncpg==0.29.0
alembic==1.13.2

# opsiyonel: supabase client (admin/batch operasyonları için)
supabase==2.6.0